refactor: setup cogshop v2 model
